deleted the raise issue microservice and replaced with new one(issue pages)
