+++ issuePages/src/main/java/com/issuePages/IssuePagesApplication.java
@@ -1,0 +1,15 @@
+package com.issuePages;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class IssuePagesApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(IssuePagesApplication.class, args);
+	}
+
+}
