+++ issuePages/src/main/java/com/issuePages/Repo/CommentRepository.java
@@ -1,0 +1,17 @@
+package com.issuePages.Repo;
+
+
+import java.util.List;
+
+//import javax.xml.stream.events.Comment;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.issuePages.Entity.Comment;
+
+@Repository
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+    List<Comment> findByIssueId(Long issueId);
+}
+
