+++ issuePages/src/main/resources/application.properties
@@ -1,9 +1,8 @@
-spring.application.name=Authentication
-
+spring.application.name=issuePages
 
 spring.datasource.url=jdbc:mysql://localhost:3306/vdi_Issue_Tracker_DataBase
 
-server.port=8586
+server.port=8585
 
 spring.datasource.username=root
 spring.datasource.password =10@Mendbotany
