+++ issuePages/src/test/java/com/issuePages/IssuePagesApplicationTests.java
@@ -1,0 +1,13 @@
+package com.issuePages;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class IssuePagesApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
