--- raiseIssue-1/.settings/org.eclipse.m2e.core.prefs
@@ -1,4 +1,0 @@
-activeProfiles=pom.xml
-eclipse.preferences.version=1
-resolveWorkspaceProjects=true
-version=1
