--- raiseIssue-1/src/main/java/test/com/demo/RaiseIssue1Application.java
@@ -1,19 +1,0 @@
-package test.com.demo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-
-@SpringBootApplication
-@EnableDiscoveryClient
-
-public class RaiseIssue1Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(RaiseIssue1Application.class, args);
-	}
-
-}
-
-
