--- raiseIssue-1/src/main/resources/application.properties
@@ -1,17 +1,0 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/vdi_Issue_Tracker_DataBase
-
-server.port=8585
-
-spring.datasource.username=root
-spring.datasource.password =root
-spring.datasource.testWhileIdle=true
-spring.datasource.validationQuery=SELECT 1
-spring.jpa.show-sql=true
-spring.jpa.hibernate.naming-strategy=org.hibernate.cfg.ImprovedNamingStrategy
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
-spring.main.allow-circular-references=true
-
-#spring.jpa.hibernate.ddl-auto=create
-spring.jpa.hibernate.ddl-auto=none
-
-eureka.client.service-url.defaultZone=http://localhost:8761/eureka/
