--- raiseIssue-1/src/test/java/test/com/demo/RaiseIssue1ApplicationTests.java
@@ -1,13 +1,0 @@
-package test.com.demo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RaiseIssue1ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
